refactor: split the objects into seperate files
