+++ aws_lambda_powertools/utilities/typing/lambda_client_context_mobile_client.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+class LambdaClientContextMobileClient(object):
+    installation_id: str
+    app_title: str
+    app_version_name: str
+    app_version_code: str
+    app_package_name: str
