+++ aws_lambda_powertools/utilities/typing/lambda_cognito_identity.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+
+class LambdaCognitoIdentity(object):
+    cognito_identity_id: str
+    cognito_identity_pool_id: str
