+++ aws_lambda_powertools/utilities/typing/lambda_dict.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from typing import Any, Dict
+
+LambdaDict = Dict[str, Any]
